Move `FusionInterruptedException` into `embed` package
This is a public feature used by the embedding API.
