+++ runtime/src/main/java/dev/ionfusion/fusion/StandardRuntime.java
@@ -16,6 +16,7 @@ import com.amazon.ion.system.IonSystemBuilder;
 import dev.ionfusion.runtime._private.cover.CoverageCollector;
 import dev.ionfusion.runtime.base.FusionException;
 import dev.ionfusion.runtime.base.ModuleIdentity;
+import dev.ionfusion.runtime.embed.FusionInterruptedException;
 import dev.ionfusion.runtime.embed.FusionRuntime;
 import dev.ionfusion.runtime.embed.ModuleBuilder;
 import dev.ionfusion.runtime.embed.SandboxBuilder;
