+++ runtime/src/main/java/dev/ionfusion/runtime/embed/TopLevel.java
@@ -5,7 +5,6 @@ package dev.ionfusion.runtime.embed;
 
 import com.amazon.ion.IonReader;
 import com.amazon.ion.IonWriter;
-import dev.ionfusion.fusion.FusionInterruptedException;
 import dev.ionfusion.runtime.base.FusionException;
 import dev.ionfusion.runtime.base.ResourceDescriptor;
 import java.io.File;
