+++ runtime/src/test/java/dev/ionfusion/fusion/InterruptionTest.java
@@ -9,6 +9,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import com.amazon.ion.IonReader;
+import dev.ionfusion.runtime.embed.FusionInterruptedException;
 import dev.ionfusion.runtime.embed.ModuleBuilder;
 import dev.ionfusion.runtime.embed.TopLevel;
 import java.util.concurrent.CountDownLatch;
